# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Atlas Platform API
  version: 0.2.0
  description: Platform operations — orgs, onboarding, webhooks, events, reports, exports, and cron jobs
  contact:
    name: AtlasPayments Team
    url: https://github.com/shivemind/atlaspayments
servers:
  - url: http://localhost:3000
    description: Local development server. Override via Postman {{baseUrl}} environment variable.
...
paths:
...
  /api/v1/webhook_endpoints/{id}:
    get:
      summary: Retrieve a webhook endpoint
      description: Returns the webhook endpoint with its secret and configuration.
      x-platform:
        kind: resource
        entity: WebhookEndpoint
        affectsLedger: false
      operationId: getWebhookEndpoint
      tags:
        - Webhooks
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Webhook endpoint details.
          content:
            application/json:
              schema:
                type: object
                properties:
                  webhook_endpoint:
                    $ref: "#/components/schemas/WebhookEndpoint"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Webhook endpoint not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      summary: Update a webhook endpoint
      description: |
        Update the URL, enabled events, or active status of a webhook endpoint.
      x-platform:
        kind: transition
        entity: WebhookEndpoint
        stateChanges:
          - updated
        affectsLedger: false
      operationId: updateWebhookEndpoint
      tags:
        - Webhooks
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  format: uri
                enabled_events:
                  type: array
                  items:
                    type: string
                is_active:
                  type: boolean
            examples:
              updateUrl:
                value:
                  url: https://example.com/new-webhook
              deactivate:
                value:
                  is_active: false
      responses:
        "200":
          description: Webhook endpoint updated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  webhook_endpoint:
                    $ref: "#/components/schemas/WebhookEndpoint"
        "400":
          description: Invalid payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Webhook endpoint not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Deactivate a webhook endpoint
      description: |
        Soft-deletes a webhook endpoint by setting is_active to false. The endpoint is retained for audit history.
      x-platform:
        kind: transition
        entity: WebhookEndpoint
        stateChanges:
          - deactivated
        affectsLedger: false
      operationId: deleteWebhookEndpoint
      tags:
        - Webhooks
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Webhook endpoint deactivated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  webhook_endpoint:
                    $ref: "#/components/schemas/WebhookEndpoint"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Webhook endpoint not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/internal/underwriting/cases:
    get:
      summary: List underwriting cases
      description: Returns a paginated list of underwriting cases with optional filters.
      x-platform:
        kind: resource
        entity: UnderwritingCase
        affectsLedger: false
      operationId: listUnderwritingCases
      tags:
        - Underwriting
      security:
        - PlatformAdminKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - in: query
          name: status
          description: Filter by case status.
          schema:
            type: string
        - in: query
          name: onboarding_profile_id
          description: Filter by onboarding profile.
          schema:
            type: string
      responses:
        "200":
          description: Paginated underwriting cases.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - pagination
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/UnderwritingCase"
                  pagination:
                    $ref: "#/components/schemas/PaginationMeta"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Create underwriting case
      description: |
        Opens a new underwriting case linked to an onboarding profile for manual or automated risk review.
      x-platform:
        kind: resource
        entity: UnderwritingCase
        emitsEvents:
          - underwriting.case.created
        affectsLedger: false
        workflowRefs:
          - onboarding_full
      operationId: createUnderwritingCase
      tags:
        - Underwriting
      security:
        - PlatformAdminKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - onboarding_profile_id
              properties:
                onboarding_profile_id:
                  type: string
                assigned_to:
                  type: string
                risk_score:
                  type: integer
                  minimum: 0
                  maximum: 1000
                notes:
                  type: string
            examples:
              basic:
                value:
                  onboarding_profile_id: cmf1profile123
                  assigned_to: analyst@example.com
                  risk_score: 35
      responses:
        "201":
          description: Underwriting case created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/UnderwritingCase"
        "400":
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Onboarding profile not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/internal/webhooks/endpoints:
    get:
      summary: List all webhook endpoints (admin)
      description: |
        Platform-wide webhook endpoint listing with optional merchant and active status filters.
      x-platform:
        kind: resource
        entity: WebhookEndpoint
        affectsLedger: false
      operationId: adminListWebhookEndpoints
      tags:
        - Webhooks
      security:
        - PlatformAdminKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - in: query
          name: merchant_id
          description: Filter by merchant ID.
          schema:
            type: string
        - in: query
          name: is_active
          description: Filter by active status.
          schema:
            type: string
            enum:
              - "true"
              - "false"
      responses:
        "200":
          description: Paginated list of webhook endpoints.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - pagination
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/WebhookEndpoint"
                  pagination:
                    $ref: "#/components/schemas/PaginationMeta"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Platform admin access required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    MerchantKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        Bearer API key for merchant-scoped operations. Pass as `Authorization: Bearer {{merchantApiKey}}`.
    PlatformAdminKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        Bearer API key for platform admin operations. Pass as `Authorization: Bearer {{platformAdminKey}}`.
  schemas:
    ErrorResponse:
      description: Standard error response envelope returned for 4xx/5xx responses.
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code.
            message:
              type: string
              description: Human-readable error description.
    PaginationMeta:
      description: Pagination metadata included in list responses.
      type: object
      required:
        - page
        - pageSize
        - total
        - hasMore
      properties:
        page:
          type: integer
          description: Current page number (1-indexed).
        pageSize:
          type: integer
          description: Number of results per page.
        total:
          type: integer
          description: Total number of matching records.
        hasMore:
          type: boolean
          description: Whether additional pages exist.
...
    WebhookEndpoint:
      description: A registered webhook endpoint for receiving event notifications.
      type: object
      properties:
        id:
          type: string
        merchant_id:
          type: string
        url:
          type: string
          format: uri
        is_active:
          type: boolean
        enabled_events:
          type: array
          items:
            type: string
        secret:
          type: string
          description: Only included on creation and rotate.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    UnderwritingCase:
      description: Underwriting review case linked to an onboarding profile.
      type: object
      properties:
        id:
          type: string
        onboarding_profile_id:
          type: string
        assigned_to:
          type: string
          nullable: true
        risk_score:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
        status:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
  parameters:
...
    PageParam:
      in: query
      name: page
      description: Page number (1-indexed).
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      in: query
      name: pageSize
      description: Number of results per page (max 100).
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    PathId:
      in: path
      name: id
      required: true
      description: Resource identifier.
      schema:
        type: string
...